:sparkles: Add file Pathable
